move PlatformStartup to Startup namespace
