+++ src/CtrDxEditor.Desktop/Program.cs
@@ -5,6 +5,7 @@ using Avalonia;
 
 using CtrDxEditor;
 using CtrDxEditor.Content;
+using CtrDxEditor.Startup;
 
 namespace CtrDxEditor.Desktop
 {
